IMPALA-7781: Clean up expr rewriter predicates
The expression rewriter often checks if an expression is null or a
literal, but does so using ad-hoc "instance of" or similar checks.
Yet, the expression nodes provide predicates that do the same job.
This change uses the defined predicates in place of the ad-hoc tests.

Add new predicates for the most common cases: literal, null literal,
null value, non-null literal, and literal-value. Converted several Expr
class methods to predicates.

The predicates handle cases that the ad-hoc tests do not. For example,
when the constant folding rule produces a NULL result, it casts that
null to some type: CAST(NULL AS INT), say. Later, other rules want to
check if that result is null. A simple check for a null literal won't
work, we need the IS_NULL_VALUE predicate which checks for the CAST
case.

This change adds predicate and revises the rewrite rules to use the
new predicates.

Also adds a number of minor Java cleanups found while doing the fix.

One of which is the toSql() for numeric values: tests showed that
the form of zero printed depended on the precision and scale, with one
value printing as 0E-38. Modified the code to emit "0" for all zero
values (since, unlike Java, SQL attaches no meaning to 0.0 as oppposed
to just 0.)

Detailed testing revealed missed simplification opportunites, especially
in CASE. Added the missing logic. For CASE, this meant a wholesale
refactoring of the code in order to better deal with nuances such as

CASE null WHEN ... END --> NULL
CASE ... ELSE NULL END --> CASE ... END
CASE ... WHEN ... THEN NULL END CASE ... END

The constant folding rule always wrapped the result in a cast,
even if the result was of the correct type. Skipped the unnecessary
cast to avoid repeated passes through this rule.

Testing:

* Refactored conditional tests into a new class, broken down
by type expression type for easier debugging.
* Created a base class for the existing and new expression test
classes.
* Created a test fixture to manage analysis tests to make it easier
to grab and inspect various aspects of the plan, as well as to control
query and other options.
* Added new tests that failed without these changes, but which pass
with them.
* Some of the plans in PlannerTest changed: verified that the new
plans differ only in the added simplifications.

Change-Id: I92d117145e427fcc5c71299096cf0e2d964ab63c