IMPALA-7807: Analysis test fixture
Refactors the existing ExprRewriteRulesTest to pull out the framework
functionality into a "test fixture" class which can be used for a
variety of tests. The fixture allows more variation, and access to
more intermediate state, than does the existing function-based
framework. For example, the fixture allows setting per-query
options, allows doing full or partial (no rewrite) analysis, provides
access to things like the Analyzer, and more.

The fixture pulls in code from FrontEndBase to handle the low-level
tasks of running a query, and from ExprRewriteRulesTest for the
rewrite-specific aspects. For now, the FrontEndBase base class was
left unchanged.  ExprRewriteRulesTest is refactored to use the new
fixture in order to illustrate usage of the test fixture.

The key value of this work is to allow greater detail when testing
future change requests.  In order to keep the refactoring as simple
and clean, no new test cases were added here

Testing: since this is a test fixture, the refactored ExprRewriteRulesTest
implicitly tests the fixture code. No "production" (non-test) code was
changed in this patch.

Change-Id: Id69c99b284960f16394f61072d38dd00269bc10c