IMPALA-7895: Incorrect expected results for spillable-buffer-sizing.test
Expected results from spillable-bufe-sizing.test and max-row-size.test
included incorrect expressions: the un-anayzed GROUP BY expression with
ordinals represented as (invalid) casts.

SelectStmt is a bit of a mess. There are two copies of the grouping
expressions. Here we want to use the analyzed version with the ordinals
replaced.

Testing:

* Problem found when running PlannerTest. PlannerTest now passes,
  with correct results, after this change.
* Turns out there is another path used when generating SQL for a view
  which does toSql() on an unanalyzed query. Added unit tests for this
  case.

Change-Id: I413bded920e27fe9f41f0ea989696a0c8f92fe4a