IMPALA-7914: Base class for statement-like AST nodes
In order to integrate expression rewrites into the analysis phase, the
expression analyze() operation must be able to replace one expression
node with another. Statements, however, are analyzed in place. The two
types of parse nodes thus need different analyze() semantics.

To prepare for that goal, this patch introduces a new StmtNode class
as the base for all statement-like AST nodes. The existing analyze()
method moves to StmtNode. While Expr still defines this method for now,
the future goal is to change the Expr analyze() semantics.

Tests: This is purely a code restructuring, no functional changes. Reran
all FE tests.

Change-Id: Ie565ff02ad74f805a667017ba9bc8c0a2697a97b