IMPALA-7927: Enhance Rewritten SQL in test files
Makes two changes to the rewritten SQL shown in .test files:

1. Show the rewritten HAVING clause. (We have two copies. Before, the
   rewritten SQL displayed the unrewritten copy.)

2. Smap substitution replaces expressions with references to
   materialized slots. This replacement was hidden in the rewritten SQL.
   This change makes the replacement explicit.

Example of HAVING with the above changes:

HAVING $ao$1 /* sum(2 + id) */ > CAST(1 AS BIGINT)

The $ao$1 is the reference to the Aggregation Output slot 1. The column
label (an expression here) is shown in a quote for reference.

More detail in the JIRA ticket.

Testing: Affects only .test files. Reran all FE tests and updated .test
files as needed.

Change-Id: I09ee971a7797c4154bafee5c080ac3c6a1057fc7