IMPALA-7968, Part 1: JSON serialization framework
Provides a set of interfaces and classes to serialize the analyzer AST
to JSON. This part provides the serialization framework itself, later
parts will provide the serialization code and test framework.

The framework is modular with three implementations:

* Jackson streaming API to produce standard JSON for export to external
  systems.
* Bespoke JSON serializer optimized to produce compact, repeatable
  output for testing. This form may be adapted to produce YAML, which is
  more compact than JSON in order to reduce test file size.
* JSONSimple (already an Impala dependency) to create JSON objects. Not
  super useful, but shows how to use other object-based JSON frameworks
  if needed.

This framework itself is independent of the AST and can also be used to
serialize the physical plan as well.

Testing: test cases exercise the mechanisms. The code is an island in
the patch; no other code uses it yet.

Change-Id: Ie3101c2708bf6cf4bec61af83a5db9b6d70ddd9c