IMPALA-7968, Part 1: JSON serialization framework
Provides a set of interfaces and classes to serialize the analyzer AST
to JSON. This part provides the serialization framework itself, later
parts will provide the serialization code and test framework.

The framework uses Jackson streaming API to JSON. Though the primary
purpose is to generate JSON for in PlannerTest text comparisons, the use
of Jackson ensures that the JSON could, if wanted, be exported to other
tools.

While Jackson provides low-level serialization, the classes here provide
a set of higher-level abstractions that make the AST serialization (not
in this patch) simpler than it would be using the low-level Jackson
tools.

This framework itself is independent of the AST and can also be used to
serialize the physical plan as well.

Testing: test cases exercise the mechanisms. The code is an island in
the patch; no other code uses it yet.

Change-Id: Ie3101c2708bf6cf4bec61af83a5db9b6d70ddd9c