IMPALA-8015: Incorrect cardinality calculation for the generic case
Impala uses two techniques to estimate join cardinality: one for FK/PK,
another for the "generic" case. (IMPALA-8018 suggests we combine the
two. But, for the purposes of this fix, we leave them separate.)

The code for the generic estimation is mostly right for a key that
consists of a singe column. But, if a key is compound (contains more
than one column), the calculations can produce incorrect results.

The JIRA ticket works out the detailed math. The code is more complex
than the math as it must also handle errors in NDV, in scan cardinality
estimation and in base table cardinality estimation.

Tests: Unfortunately, the test suite contains no tables that use
compound keys. Existing tests simulate the sutuation by using highly
correlated keys. E.g. a unique id along with an additional column. Such
tests are not realistic. Because adding such tables is a large project,
it is left for a separate patch.

Instead, testing was done using a production query that misfired without
this fix, but that produced correct results with this fix.

Change-Id: I32bb364931dfecc00ff3ae3989b76a091cdcdbe2